Feat: #141 - 파일분리 및 리팩토링
