+++ 2024-MacC-M4-6princess/2024-MacC-M4-6princess/Source/View/ImageOutput/IOView.swift
@@ -85,23 +85,13 @@ struct IOView: View {
                             }
                             Button(action: {
                                 
-                                viewModel.ShowShare.toggle()
-                                //                                viewModel.showAcitivity.toggle()
+                                viewModel.showShareButton.toggle()
                             }) {
                                 Image("share.icon")
                                     .resizable()
                                     .frame(width:60,height: 60)
                                 
-                                //                                Rectangle()
-                                //                                    .foregroundColor(.clear)
-                                //                                    .frame(height: 60)
-                                //                                    .background(Color.pointPink)
-                                //                                    .cornerRadius(10)
-                                //                                    .overlay(
-                                //                                        Text("SNS 공유")
-                                //                                            .foregroundColor(.white)
-                                //                                            .font(.system(size: 18, weight: .bold))
-                                //                                    )
+
                             }
                         }
                         .frame(maxWidth: .infinity)
@@ -127,7 +117,6 @@ struct IOView: View {
                                     .frame(height: 40)
                                     .background(Color.pointPink)
                                     .cornerRadius(10)
-                                //                                    .padding(.trailing)
                                     .overlay(
                                         Text("카메라로 이동")
                                             .foregroundColor(.white)
@@ -135,24 +124,12 @@ struct IOView: View {
                                     )
                             }
                             Button(action: {
-                                viewModel.ShowShare = true
-                                //                                viewModel.showAcitivity.toggle()
+                                viewModel.showShareButton = true
                             }
                             ) {
                                 Image("share.icon")
                                     .resizable()
                                     .frame(width:40,height: 40)
-                                
-                                //                                Rectangle()
-                                //                                    .foregroundColor(.clear)
-                                //                                    .frame(height: 40)
-                                //                                    .background(Color.pointPink)
-                                //                                    .cornerRadius(10)
-                                //                                    .overlay(
-                                //                                        Text("SNS 공유")
-                                //                                            .foregroundColor(.white)
-                                //                                            .font(.system(size: 18, weight: .bold))
-                                //                                    )
                             }
                         }
                         .frame(maxWidth: .infinity)
@@ -175,12 +152,12 @@ struct IOView: View {
         .alert(isPresented: $viewModel.showAlert) {
             Alert(title: Text("오류 발생"), message: Text(viewModel.alertMessage), dismissButton: .default(Text("확인")))
         }
-        .sheet(isPresented: $viewModel.ShowShare) {
+        .sheet(isPresented: $viewModel.showShareButton) {
             BottomSheetViewWrapper(viewModel: viewModel)
                 .presentationDetents([.height(150)])
         }
-        .onChange(of: viewModel.ShowShare, perform: { newValue in
-            if viewModel.ShowShare == false && viewModel.showAcitivity == true{
+        .onChange(of: viewModel.showShareButton, perform: { newValue in
+            if viewModel.showShareButton == false && viewModel.showAcitivity == true{
                 print("onchange start")
                 DispatchQueue.main.asyncAfter(deadline: .now() + 0.1){
                     if let _ = viewModel.compositeImage {
@@ -190,202 +167,20 @@ struct IOView: View {
                 }
             }
         })
-        //        .onChange(of: viewModel.ShowShare) { newValue in
-        //            if !newValue {
-        //                DispatchQueue.main.async {
-        //                    if let _ = viewModel.compositeImage {
-        //                        viewModel.showAcitivity = true
-        //                    }
-        //                }
-        //            }
-        //        }
-        
-        //        .background(
-        //            if viewModel.ShowShare{
-        //                BottomSheetViewWrapper(viewModel: viewModel)
-        ////                    .frame(height:150)
-        //            }
-        //            else if viewModel.showAcitivity{
-        //                Group{
-        //                    if let photo = viewModel.compositeImage
-        //                        ,viewModel.showAcitivity
-        //                    {
-        //                        ShareSheet(isPresented: $viewModel.showAcitivity, shareData: (photo,"title","Frameet으로 사진 낋여왔음"))
-        //                    }
-        //                    
-        //                }
-        //            }
-        //           
-        //        )
         .overlay(
             Group {
-                //            if viewModel.ShowShare {
-                //                BottomSheetViewWrapper(viewModel: viewModel)
-                //                    .onDisappear {
-                //                        DispatchQueue.main.asyncAfter(deadline: .now() + 0.05) {
-                //                            if viewModel.showAcitivity{
-                //                                viewModel.changeOverlay = true
-                //                            }
-                //                        }
-                //                    }
-                //            }
-                //            else
+
                 if viewModel.changeOverlay, let photo = viewModel.compositeImage {
-                    ShareSheet(isPresented: $viewModel.changeOverlay, shareData: (photo, "title", "Frameet으로 사진 찍어왔음"))
+                    IOShareSheet(isPresented: $viewModel.changeOverlay, shareData: (photo, "title", "Frameet으로 사진 찍어왔음"))
                         .onAppear{
                             print("sharesheet start")
                         }
                 }
             }
         )
-        
-        
-        // 상단 툴바
         .navigationBarBackButtonHidden()
     }
     var canvasView: some View {
         IOCanvasView(viewModel: viewModel)
     }
-}
-
-import UIKit
-import LinkPresentation
-
-struct ShareSheet: UIViewControllerRepresentable {
-    @Binding var isPresented: Bool
-    var shareData: (image: UIImage, title: String, content: String)
-    
-    func makeUIViewController(context: Context) -> UIViewController {
-        UIViewController()
-    }
-    
-    func updateUIViewController(_ uiViewController: UIViewController, context: Context) {
-        if isPresented && uiViewController.presentedViewController == nil {
-            let items: [Any] = [
-                SharePinNumberActivityItemSource(title: shareData.title, content: shareData.content, photo: shareData.image)
-            ]
-            
-            let activityVC = UIActivityViewController(activityItems: items, applicationActivities: nil)
-            activityVC.completionWithItemsHandler = { _, _, _, _ in
-                DispatchQueue.main.async {
-                    self.isPresented = false
-                }
-            }
-            uiViewController.present(activityVC, animated: true, completion: nil)
-        }
-    }
-}
-
-final class SharePinNumberActivityItemSource: NSObject, UIActivityItemSource {
-    private var title: String
-    private var content: String
-    private var image: UIImage
-    
-    init(title: String, content: String, photo: UIImage) {
-        self.title = title
-        self.content = content
-        self.image = photo
-        super.init()
-    }
-    
-    func activityViewControllerPlaceholderItem(_ activityViewController: UIActivityViewController) -> Any {
-        return content
-    }
-    
-    func activityViewController(_ activityViewController: UIActivityViewController, itemForActivityType activityType: UIActivity.ActivityType?) -> Any? {
-        // PNG 데이터로 변환
-        guard let pngData = image.pngData() else { return content }
-                
-//                if activityType == .airDrop {
-//                    return pngData
-//                }
-                
-                // 최대 크기 설정 (Twitter: 4096x4096, Instagram: 1080x1080)
-                let maxSize: CGFloat = 2096
-                let instagramMaxSize: CGFloat = 1080
-                let imageSize = image.size
-                let aspectRatio = imageSize.width / imageSize.height
-                
-                var newWidth = imageSize.width
-                var newHeight = imageSize.height
-                
-                // Twitter 크기 제한 적용
-                if imageSize.width > maxSize || imageSize.height > maxSize {
-                    if imageSize.width > imageSize.height {
-                        newWidth = maxSize
-                        newHeight = newWidth / aspectRatio
-                    } else {
-                        
-                        newHeight = maxSize
-                        newWidth = newHeight * aspectRatio
-                    }
-                }
-                
-                // Instagram 크기 제한 적용 (더 엄격한 조건)
-                if newWidth > instagramMaxSize || newHeight > instagramMaxSize {
-                    if newWidth > newHeight {
-                        newWidth = instagramMaxSize
-                        newHeight = newWidth / aspectRatio
-                    } else {
-                        newHeight = instagramMaxSize
-                        newWidth = newHeight * aspectRatio
-                    }
-                }
-                
-                // 리사이즈된 이미지 생성
-                let newSize = CGSize(width: newWidth, height: newHeight)
-                UIGraphicsBeginImageContextWithOptions(newSize, true, image.scale) // opaque를 true로 설정
-                image.draw(in: CGRect(origin: .zero, size: newSize))
-                let resizedImage = UIGraphicsGetImageFromCurrentImageContext()
-                UIGraphicsEndImageContext()
-                
-                // 리사이즈된 이미지를 PNG로 변환 및 크기 확인
-                if let resizedImageData = resizedImage?.pngData() {
-                    if resizedImageData.count <= 30_000_000 { // 30MB 이하
-                        return resizedImageData
-                    } else {
-                        print("Image size exceeds 30MB, resizing further might be required.")
-                        // 추가 압축 또는 크기 조정 로직을 여기에 구현 가능
-                    }
-                }
-                
-                return content
-        
-    }
-    
-    func activityViewController(_ activityViewController: UIActivityViewController, subjectForActivityType activityType: UIActivity.ActivityType?) -> String {
-        return title
-    }
-    
-    func activityViewControllerLinkMetadata(_ activityViewController: UIActivityViewController) -> LPLinkMetadata? {
-        let metaData = LPLinkMetadata()
-        metaData.title = content
-        metaData.iconProvider = NSItemProvider(object: image)
-        metaData.originalURL = URL(string: "https://apps.apple.com/kr/app/frameet-%ED%94%84%EB%A0%88%EC%9E%84%EB%B0%8B-%EC%B5%9C%EC%95%A0%EC%99%80-%ED%95%A8%EA%BB%98-%ED%8A%B9%EB%B3%84%ED%95%9C-%EC%9D%BC%EC%83%81/id6737822930")
-        return metaData
-    }
-}
-//        .sheet(isPresented: $viewModel.showAcitivity){
-////            BottomSheetView()
-////                .presentationDetents([.height(300)])
-////            IOSNSView(viewModel: viewModel)
-////                .presentationDetents([.height(300)])
-//            Group{
-//                if let photo = viewModel.compositeImage
-//                {
-//                    ShareSheet(isPresented: $viewModel.showAcitivity, shareData: (photo,"title","Frameet으로 사진 낋여왔음"))
-//                        .presentationDetents([.height(300)])
-//                }
-//            }
-//        }
-import SwiftUI
-
-struct BottomSheetViewWrapper: UIViewControllerRepresentable {
-    var viewModel: IOViewModel
-    
-    func makeUIViewController(context: Context) -> BottomSheetViewController {
-        return BottomSheetViewController(viewModel: viewModel)
-    }
-    
-    func updateUIViewController(_ uiViewController: BottomSheetViewController, context: Context) {}
 }
